+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 ### Changed
 
 - Bump uniffi to 0.29.4 (https://github.com/rust-nostr/nostr-sdk-ffi/pull/49)
-- Bump nostr from 0.43.0 to 19970a45 (see the Upstream CHANGELOG for more details)
+- Bump nostr from 0.43.0 to 0.44.0 (see the Upstream CHANGELOG for more details)
 
 ## v0.43.0 - 2025/07/28